rename task files
