+++ src/components/VerseCard.tsx
